Move header files out of `src` when they are used in other pacakges
This fixes more compilation erros that I have when building
this package on Arch Linux.
